waiting events
event has expanded dramatically --
now the behaviour for a 'sleeping' event has been added and is contained within the event object. in this time the event will do nothing until it has counted its wait down.

additionally, whether a character is present is now a field of that person object rather than a separate array (as makes sense).